# api/src/generated/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"DefaultSelection__36_VenuePayload_": {
				"properties": {
					"deleted_at": {
						"type": "string",
						"format": "date-time"
					},
					"updated_at": {
						"type": "string",
						"format": "date-time"
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"is_active": {
						"type": "boolean"
					},
					"address": {
						"type": "string"
					},
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"deleted_at",
					"updated_at",
					"created_at",
					"is_active",
					"address",
					"id",
					"name"
				],
				"type": "object"
			},
			"VenueModel": {
				"$ref": "#/components/schemas/DefaultSelection__36_VenuePayload_",
				"description": "Model Venue"
			},
			"Venue": {
				"$ref": "#/components/schemas/VenueModel",
				"description": "Model Venue"
			},
			"PaginationMeta": {
				"properties": {
					"total": {
						"type": "number",
						"format": "double"
					},
					"page": {
						"type": "number",
						"format": "double"
					},
					"per_page": {
						"type": "number",
						"format": "double"
					},
					"last_page": {
						"type": "number",
						"format": "double"
					},
					"has_next": {
						"type": "boolean"
					}
				},
				"required": [
					"total",
					"page",
					"per_page",
					"last_page",
					"has_next"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PaginatedResult_Venue_": {
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/Venue"
						},
						"type": "array"
					},
					"meta": {
						"$ref": "#/components/schemas/PaginationMeta"
					}
				},
				"required": [
					"data",
					"meta"
				],
				"type": "object",
				"additionalProperties": false
			},
			"infer_typeofcreateVenueSchema_": {
				"properties": {
					"address": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"address",
					"name"
				],
				"type": "object"
			},
			"CreateVenueDto": {
				"$ref": "#/components/schemas/infer_typeofcreateVenueSchema_"
			},
			"infer_typeofupdateVenueSchema_": {
				"properties": {
					"is_active": {
						"type": "boolean"
					},
					"address": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"type": "object"
			},
			"UpdateVenueDto": {
				"$ref": "#/components/schemas/infer_typeofupdateVenueSchema_"
			},
			"Pick_User.Exclude_keyofUser.password__": {
				"properties": {
					"name": {
						"type": "string"
					},
					"id": {
						"type": "number",
						"format": "double"
					},
					"is_active": {
						"type": "boolean"
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"updated_at": {
						"type": "string",
						"format": "date-time"
					},
					"email": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"email_verified": {
						"type": "boolean"
					},
					"email_verified_at": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"name",
					"id",
					"is_active",
					"created_at",
					"updated_at",
					"email",
					"phone",
					"email_verified",
					"email_verified_at"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_User.password_": {
				"$ref": "#/components/schemas/Pick_User.Exclude_keyofUser.password__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"SafeUser": {
				"$ref": "#/components/schemas/Omit_User.password_"
			},
...
			"infer_typeofupdateUserSchema_": {
				"properties": {
					"role_ids": {
						"items": {
							"type": "number",
							"format": "double"
						},
						"type": "array"
					},
					"phone": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"type": "object"
			},
			"UpdateUserDto": {
				"$ref": "#/components/schemas/infer_typeofupdateUserSchema_"
			},
...
			"TiebreakerOption": {
				"type": "string",
				"enum": [
					"goal_diff",
					"goals_scored",
					"head_to_head",
					"goals_conceded",
					"yellow_cards",
					"red_cards"
				]
			},
			"TournamentRuleDto": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"tournament_id": {
						"type": "number",
						"format": "double"
					},
					"is_active": {
						"type": "boolean"
					},
					"points_per_win": {
						"type": "number",
						"format": "double"
					},
					"points_per_draw": {
						"type": "number",
						"format": "double"
					},
					"points_per_loss": {
						"type": "number",
						"format": "double"
					},
					"forfeit_score": {
						"type": "number",
						"format": "double"
					},
					"yellow_cards_suspension": {
						"type": "number",
						"format": "double"
					},
					"max_players_per_team": {
						"type": "number",
						"format": "double"
					},
					"min_players_per_team": {
						"type": "number",
						"format": "double"
					},
					"teams_advance_per_group": {
						"type": "number",
						"format": "double"
					},
					"tiebreaker_order": {
						"items": {
							"$ref": "#/components/schemas/TiebreakerOption"
						},
						"type": "array"
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"updated_at": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"user": {
						"properties": {
							"phone": {
								"type": "string",
								"nullable": true
							},
							"email": {
								"type": "string"
							},
							"name": {
								"type": "string"
							},
							"id": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"phone",
							"email",
							"name",
							"id"
						],
						"type": "object",
						"nullable": true
					},
					"tournament": {
						"properties": {
							"name": {
								"type": "string"
							},
							"id": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"name",
							"id"
						],
						"type": "object",
						"nullable": true
					}
				},
				"required": [
					"id",
					"tournament_id",
					"is_active",
					"points_per_win",
					"points_per_draw",
					"points_per_loss",
					"forfeit_score",
					"yellow_cards_suspension",
					"max_players_per_team",
					"min_players_per_team",
					"teams_advance_per_group",
					"tiebreaker_order",
					"created_at",
					"updated_at"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"infer_typeofupdateTournamentRuleSchema_": {
				"properties": {
					"tiebreaker_order": {
						"items": {
							"type": "string",
							"enum": [
								"goal_diff",
								"goals_scored",
								"head_to_head",
								"goals_conceded",
								"yellow_cards",
								"red_cards"
							]
						},
						"type": "array"
					},
					"teams_advance_per_group": {
						"type": "number",
						"format": "double"
					},
					"min_players_per_team": {
						"type": "number",
						"format": "double"
					},
					"max_players_per_team": {
						"type": "number",
						"format": "double"
					},
					"yellow_cards_suspension": {
						"type": "number",
						"format": "double"
					},
					"forfeit_score": {
						"type": "number",
						"format": "double"
					},
					"points_per_loss": {
						"type": "number",
						"format": "double"
					},
					"points_per_draw": {
						"type": "number",
						"format": "double"
					},
					"points_per_win": {
						"type": "number",
						"format": "double"
					},
					"tournament_id": {
						"type": "number",
						"format": "double"
					}
				},
				"type": "object"
			},
			"UpdateTournamentRuleDto": {
				"$ref": "#/components/schemas/infer_typeofupdateTournamentRuleSchema_"
			},
			"DefaultSelection__36_TournamentPayload_": {
				"properties": {
					"user_id": {
						"type": "number",
						"format": "double"
					},
					"logo": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"deleted_at": {
						"type": "string",
						"format": "date-time"
					},
					"updated_at": {
						"type": "string",
						"format": "date-time"
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"is_active": {
						"type": "boolean"
					},
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"user_id",
					"logo",
					"description",
					"deleted_at",
					"updated_at",
					"created_at",
					"is_active",
					"id",
					"name"
				],
				"type": "object"
			},
			"TournamentModel": {
				"$ref": "#/components/schemas/DefaultSelection__36_TournamentPayload_",
				"description": "Model Tournament"
			},
			"Tournament": {
				"$ref": "#/components/schemas/TournamentModel",
				"description": "Model Tournament"
			},
...
			"infer_typeofupdateTournamentSchema_": {
				"properties": {
					"is_active": {
						"type": "boolean"
					},
					"logo": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"type": "object"
			},
			"UpdateTournamentDto": {
				"$ref": "#/components/schemas/infer_typeofupdateTournamentSchema_"
			},
			"DefaultSelection__36_TeamPayload_": {
				"properties": {
					"coach_name": {
						"type": "string"
					},
					"user_id": {
						"type": "number",
						"format": "double"
					},
					"logo": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"deleted_at": {
						"type": "string",
						"format": "date-time"
					},
					"updated_at": {
						"type": "string",
						"format": "date-time"
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"is_active": {
						"type": "boolean"
					},
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"coach_name",
					"user_id",
					"logo",
					"description",
					"deleted_at",
					"updated_at",
					"created_at",
					"is_active",
					"id",
					"name"
				],
				"type": "object"
			},
			"TeamModel": {
				"$ref": "#/components/schemas/DefaultSelection__36_TeamPayload_",
				"description": "Model Team"
			},
			"Team": {
				"$ref": "#/components/schemas/TeamModel",
				"description": "Model Team"
			},
...
			"DefaultSelection__36_SeasonTeamPayload_": {
				"properties": {
					"group_id": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"$ref": "#/components/schemas/SeasonTeamStatus"
					},
					"season_id": {
						"type": "number",
						"format": "double"
					},
					"team_id": {
						"type": "number",
						"format": "double"
					},
					"user_id": {
						"type": "number",
						"format": "double"
					},
					"deleted_at": {
						"type": "string",
						"format": "date-time"
					},
					"updated_at": {
						"type": "string",
						"format": "date-time"
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"is_active": {
						"type": "boolean"
					},
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"group_id",
					"status",
					"season_id",
					"team_id",
					"user_id",
					"deleted_at",
					"updated_at",
					"created_at",
					"is_active",
					"id"
				],
				"type": "object"
			},
			"SeasonTeamModel": {
				"$ref": "#/components/schemas/DefaultSelection__36_SeasonTeamPayload_",
				"description": "Model SeasonTeam"
			},
			"SeasonTeam": {
				"$ref": "#/components/schemas/SeasonTeamModel",
				"description": "Model SeasonTeam"
			},
...
			"SeasonModel": {
				"$ref": "#/components/schemas/DefaultSelection__36_SeasonPayload_",
				"description": "Model Season"
			},
			"Season": {
				"$ref": "#/components/schemas/SeasonModel",
				"description": "Model Season"
			},
...
			"infer_typeofupdateSeasonSchema_": {
				"properties": {
					"is_registration_open": {
						"type": "boolean"
					},
					"max_teams": {
						"type": "number",
						"format": "double"
					},
					"registration_deadline": {
						"type": "string",
						"format": "date-time"
					},
					"end_date": {
						"type": "string",
						"format": "date-time"
					},
					"start_date": {
						"type": "string",
						"format": "date-time"
					},
					"status": {
						"type": "string",
						"enum": [
							"upcoming",
							"registration_open",
							"ongoing",
							"finished",
							"cancelled"
						]
					},
					"description": {
						"type": "string"
					},
					"is_active": {
						"type": "boolean"
					},
					"name": {
						"type": "string"
					}
				},
				"type": "object"
			},
			"UpdateSeasonDto": {
				"$ref": "#/components/schemas/infer_typeofupdateSeasonSchema_"
			},
...
			"DefaultSelection__36_RolePayload_": {
				"properties": {
					"description": {
						"type": "string"
					},
					"updated_at": {
						"type": "string",
						"format": "date-time"
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"is_active": {
						"type": "boolean"
					},
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"description",
					"updated_at",
					"created_at",
					"is_active",
					"id",
					"name"
				],
				"type": "object"
			},
			"RoleModel": {
				"$ref": "#/components/schemas/DefaultSelection__36_RolePayload_",
				"description": "Model Role"
			},
			"Role": {
				"$ref": "#/components/schemas/RoleModel",
				"description": "Model Role"
			},
...
			"infer_typeofupdateRoleSchema_": {
				"properties": {
					"is_active": {
						"type": "boolean"
					},
					"description": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"type": "object"
			},
			"UpdateRoleDto": {
				"$ref": "#/components/schemas/infer_typeofupdateRoleSchema_"
			},
			"PlayerPosition": {
				"type": "string",
				"enum": [
					"goalkeeper",
					"defender",
					"midfielder",
					"forward"
				]
			},
			"PlayerDto": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"date_of_birth": {
						"type": "string",
						"format": "date-time"
					},
					"position": {
						"$ref": "#/components/schemas/PlayerPosition"
					},
					"height": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"weight": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"nationality": {
						"type": "string",
						"nullable": true
					},
					"avatar": {
						"type": "string",
						"nullable": true
					},
					"is_active": {
						"type": "boolean"
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"updated_at": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"user": {
						"properties": {
							"phone": {
								"type": "string",
								"nullable": true
							},
							"email": {
								"type": "string"
							},
							"name": {
								"type": "string"
							},
							"id": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"phone",
							"email",
							"name",
							"id"
						],
						"type": "object",
						"nullable": true
					},
					"user_id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"date_of_birth",
					"position",
					"height",
					"weight",
					"nationality",
					"avatar",
					"is_active",
					"created_at",
					"updated_at",
					"user_id"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"infer_typeofupdatePlayerSchema_": {
				"properties": {
					"avatar": {
						"type": "string"
					},
					"nationality": {
						"type": "string"
					},
					"weight": {
						"type": "number",
						"format": "double"
					},
					"height": {
						"type": "number",
						"format": "double"
					},
					"position": {
						"type": "string",
						"enum": [
							"goalkeeper",
							"defender",
							"midfielder",
							"forward"
						]
					},
					"date_of_birth": {
						"type": "string",
						"format": "date-time"
					}
				},
				"type": "object"
			},
			"UpdatePlayerDto": {
				"$ref": "#/components/schemas/infer_typeofupdatePlayerSchema_"
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "api",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/venues": {
			"get": {
				"operationId": "FindAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaginatedResult_Venue_"
								}
							}
						}
					}
				},
				"tags": [
					"Venues"
				],
				"security": [
					{
						"jwt": [
							"user",
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "per_page",
						"required": false,
						"schema": {
							"default": 20,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "q",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "sort",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "direction",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"asc",
								"desc"
							]
						}
					}
				]
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Venue"
								}
							}
						}
					}
				},
				"tags": [
					"Venues"
				],
				"security": [
					{
						"jwt": [
							"user",
							"admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateVenueDto"
							}
						}
					}
				}
			}
		},
		"/venues/{id}": {
			"get": {
				"operationId": "FindById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Venue"
								}
							}
						}
					}
				},
				"tags": [
					"Venues"
				],
				"security": [
					{
						"jwt": [
							"user",
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"patch": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Venue"
								}
							}
						}
					}
				},
				"tags": [
					"Venues"
				],
				"security": [
					{
						"jwt": [
							"user",
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateVenueDto"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "SoftDelete",
				"responses": {
					"204": {
						"description": "Deleted"
					}
				},
				"tags": [
					"Venues"
				],
				"security": [
					{
						"jwt": [
							"user",
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/users/{id}": {
			"get": {
				"operationId": "FindById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SafeUser"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"patch": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SafeUser"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateUserDto"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "SoftDelete",
				"responses": {
					"204": {
						"description": "Deleted"
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/tournamentrules/{id}": {
			"get": {
				"operationId": "FindById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TournamentRuleDto"
								}
							}
						}
					}
				},
				"tags": [
					"TournamentRules"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"patch": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UpdateTournamentRuleDto"
								}
							}
						}
					}
				},
				"tags": [
					"TournamentRules"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateTournamentRuleDto"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "SoftDelete",
				"responses": {
					"204": {
						"description": "Deleted"
					}
				},
				"tags": [
					"TournamentRules"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/tournaments/{id}": {
			"get": {
				"operationId": "FindById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Tournament"
								}
							}
						}
					}
				},
				"tags": [
					"Tournaments"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"patch": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Tournament"
								}
							}
						}
					}
				},
				"tags": [
					"Tournaments"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateTournamentDto"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "SoftDelete",
				"responses": {
					"204": {
						"description": "Deleted"
					}
				},
				"tags": [
					"Tournaments"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/teams/{id}": {
			"get": {
				"operationId": "FindById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Team"
								}
							}
						}
					}
				},
				"tags": [
					"Teams"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"patch": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Team"
								}
							}
						}
					}
				},
				"tags": [
					"Teams"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": false,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"name": {
										"type": "string"
									},
									"coach_name": {
										"type": "string"
									},
									"description": {
										"type": "string"
									},
									"logo": {
										"type": "string",
										"format": "binary"
									}
								}
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "SoftDelete",
				"responses": {
					"204": {
						"description": "Deleted"
					}
				},
				"tags": [
					"Teams"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/seasonteams/{id}": {
			"get": {
				"operationId": "FindById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SeasonTeam"
								}
							}
						}
					}
				},
				"tags": [
					"SeasonTeams"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"delete": {
				"operationId": "SoftDelete",
				"responses": {
					"204": {
						"description": "Deleted"
					}
				},
				"tags": [
					"SeasonTeams"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/seasons/{id}": {
			"get": {
				"operationId": "FindById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Season"
								}
							}
						}
					}
				},
				"tags": [
					"Seasons"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"patch": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Season"
								}
							}
						}
					}
				},
				"tags": [
					"Seasons"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateSeasonDto"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "SoftDelete",
				"responses": {
					"204": {
						"description": "Deleted"
					}
				},
				"tags": [
					"Seasons"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/roles/{id}": {
			"get": {
				"operationId": "FindById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Role"
								}
							}
						}
					}
				},
				"tags": [
					"Roles"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"patch": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Role"
								}
							}
						}
					}
				},
				"tags": [
					"Roles"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateRoleDto"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "SoftDelete",
				"responses": {
					"204": {
						"description": "Deleted"
					}
				},
				"tags": [
					"Roles"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/players/{id}": {
			"get": {
				"operationId": "FindById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PlayerDto"
								}
							}
						}
					}
				},
				"tags": [
					"Players"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"user",
							"organizing",
							"guest"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"patch": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PlayerDto"
								}
							}
						}
					}
				},
				"tags": [
					"Players"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"user",
							"organizing",
							"guest"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdatePlayerDto"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "SoftDelete",
				"responses": {
					"204": {
						"description": "Deleted"
					}
				},
				"tags": [
					"Players"
				],
				"security": [
					{
						"jwt": [
							"admin",
							"user",
							"organizing",
							"guest"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/api/v1"
		}
	]
}